+++ frontend/src/routes/tournaments/[tournamentId]/standings/+page.ts
